Agent 294 perform python-modernize automated changes (2to3migration) (#349)
* AGENT-294-Perform-automated-changes-(2to3migration):
* Apply python-modernize only.
* long to int transition.
* Make modules/scripts to work with absolute_import.
* add modernize script.
* add modernize check to circleci.

+++ scalyr_agent/json_lib/parser.py
@@ -18,6 +18,10 @@
 #
 # author: Steven Czerwinski <czerwin@scalyr.com>
 
+from __future__ import absolute_import
+from six import unichr
+from six.moves import range
+
 __author__ = "czerwin@scalyr.com"
 
 import struct
@@ -582,7 +586,7 @@ class JsonParser(object):
             and number_string.find("E") < 0
         ):
             try:
-                return sign * long(number_string)
+                return sign * int(number_string)
             except ValueError:
                 self.__error("Could not parse number as long '%s'" % number_string)
         else:
